+++ packages/widgetbook_cli/lib/src/commands/build_push.dart
@@ -19,7 +19,6 @@ class BuildPushCommand extends CliCommand<BuildPushArgs> {
     required super.context,
     this.processManager = const LocalProcessManager(),
     this.fileSystem = const LocalFileSystem(),
-    this.zipEncoder = const ZipEncoder(),
     this.cacheReader = const CacheReader(),
     WidgetbookHttpClient? cloudClient,
     StorageClient? storageClient,
@@ -70,7 +69,6 @@ class BuildPushCommand extends CliCommand<BuildPushArgs> {
   final StorageClient storageClient;
   final ProcessManager processManager;
   final FileSystem fileSystem;
-  final ZipEncoder zipEncoder;
   final CacheReader cacheReader;
 
   @override