+++ packages/widgetbook_cli/lib/src/core/cli_runner.dart
@@ -7,7 +7,6 @@ import 'package:pub_updater/pub_updater.dart';
 
 import '../../metadata.dart';
 import '../commands/cloud.dart';
-import '../commands/publish.dart';
 import '../commands/upgrade.dart';
 import '../utils/utils.dart';
 import 'context.dart';
@@ -31,13 +30,6 @@ class CliRunner extends CommandRunner<int> {
         context: context,
         logger: _logger,
         pubUpdater: _pubUpdater,
-      ),
-    );
-
-    addCommand(
-      PublishCommand(
-        context: context,
-        logger: _logger,
       ),
     );